+++ CHANGES.rst
@@ -77,6 +77,8 @@ Bug Fixes
 Cubeviz
 ^^^^^^^
 
+- Re-enable support for exporting spectrum-viewer. [#2825]
+
 Imviz
 ^^^^^